# Use batch productstatuses and vanilla WP functions