--- a/src/Jobs/CleanupProductsJob.php
+++ b/src/Jobs/CleanupProductsJob.php
@@ -36,7 +36,7 @@ class CleanupProductsJob extends AbstractProductSyncerBatchedJob {
 	 * @return array
 	 */
 	public function get_batch( int $batch_number ): array {
-		return $this->product_repository->find_synced_product_ids( $this->get_batch_size(), $this->get_query_offset( $batch_number ) );
+		return $this->product_repository->find_synced_product_ids( [], $this->get_batch_size(), $this->get_query_offset( $batch_number ) );
 	}
 
 	/**