--- a/src/Jobs/DeleteAllProducts.php
+++ b/src/Jobs/DeleteAllProducts.php
@@ -35,7 +35,7 @@ class DeleteAllProducts extends AbstractProductSyncerBatchedJob {
 	 * @return int[]
 	 */
 	protected function get_batch( int $batch_number ): array {
-		return $this->product_repository->find_synced_product_ids( $this->get_batch_size(), $this->get_query_offset( $batch_number ) );
+		return $this->product_repository->find_synced_product_ids( [], $this->get_batch_size(), $this->get_query_offset( $batch_number ) );
 	}
 
 	/**