@layoutd 2021-06-18
Thanks for the reviews @mikkamp and @nima-karimi. I'm working one a raw DB update for the products' `mc_status`, then I'll resubmit for review 🔎