@nima-karimi 2021-06-18
Just wanted to post something I tested now:

I ran a memory trace using xdebug and found the following results on which methods use the most memory when calling the `refresh` API:

```
Showing the 10 most costly calls sorted by 'memory-own'.
```

| function  | #calls | Inclusive time | Inclusive memory | Own time | Own memory |
| -------------  | ------------- | ------------- | ------------- | ------------- | ------------- |
| mysqli_query | 33540 | 10.9594 | 462801184 | 10.9594 | 462801184 |
| mysqli_fetch_object | 347837 | 5.6623 | 191927392 | 5.6623 | 191927392 |
| update_meta_cache | 9473 | 34.6720 | 275119800 | 1.8020 | 95213608 |
| preg_match_all | 24701 | 0.4039 | 53924728 | 0.4039 | 53924728 |
| in_array | 437234 | 5.9432 | 45471688 | 5.9432 | 45471688 |
| array_keys | 73406 | 1.0658 | 41758000 | 1.0658 | 41758000 |
| spl_object_hash | 129132 | 17.5383 | 37190072 | 17.5383 | 37190072 |
| WP_Post::get_instance | 59344 | 64.1508 | 139629928 | 3.5004 | 24591272 |
| array_merge | 13983 | 0.2457 | 23800752 | 0.2457 | 23800752 |
| str_replace | 179542 | 2.6249 | 17435584 | 2.6249 | 17435584 |

It seems MySQL uses the most memory here. Not sure if moving to a raw query will fix that since there would still be tons of products to update which means a lot of queries.