@mikkamp 2021-06-21
The eighth row shows that it's getting about 60k posts, `get_instance` will load the post directly from the DB if it's not available in the cache. It also shows there that `update_meta_cache` is only called 9k times.
So my conclusion would sooner be, that if the raw query avoids needing to load each post then you would get a lower memory/resource footprint.

We would need to take the meta cache into account, but that's only an issue if you read from the same postmeta within the same request. As far as I understand we mostly just query for ID's with a specific meta, which would be a query that would go directly to the DB.