@layoutd 2021-06-21
I've made a few changes, mostly to retrieve and store postmeta values with raw DB queries using [`ProductMetaQueryHelper`](https://github.com/woocommerce/google-listings-and-ads/pull/769/commits/c4fc82eb03abb378979a161ca22bdd300544680f). That way some product metadata can be retrieved as post ID-meta value pairs, to reduce memory footprint, and meta values can be batch inserted/edited to reduce the number of DB queries. 
The logic for updating individual products' MC statuses has been updated as well, so that only products whose value changes are modified (insert/update where necessary). Execution time will still depend on the number of changes to be made, but since they're batch raw quieres, the difference is not that noticeable.